fix: Resolve multiple controller type errors
- RadiologyController: Fix branchId null coalescing, TestCategory->ImagingModality mapping, TurnaroundTimeHours, FullNameEn, Items collection
- InventoryController: Cast TransactionType enums to int for comparison
- HRController: Convert TimeOnly to TimeSpan for comparison
- ClinicalVisitsController: Remove ToString() from ProviderId (already int?)

+++ XenonClinic.Api/Controllers/InventoryController.cs
@@ -656,7 +656,7 @@ public class InventoryController : BaseApiController
 
             if (request.TransactionType.HasValue)
             {
-                filteredTransactions = filteredTransactions.Where(t => t.TransactionType == request.TransactionType.Value);
+                filteredTransactions = filteredTransactions.Where(t => (int)t.TransactionType == (int)request.TransactionType.Value);
             }
 
             if (request.PatientId.HasValue)
